+++ src/kernel/mod.rs
@@ -12,6 +12,8 @@ const KALLSYMS: &str = "/proc/kallsyms";
 /// The path to the `/proc/modules` file.
 const MODULES: &str = "/proc/modules";
 
+use self::depmod::DepmodIndex;
+use self::modmap::ModMap;
 
 pub(crate) use cache::KernelCache;
 pub(crate) use resolver::KernelResolver;