+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added support for relocatable ELF and DWARF
+- Added support for kernel module symbolization with DWARF data
 - Bumped minimum supported Rust version to `1.88`